requirements.txt
astroid==2.4.1
bandit==1.6.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
emojis==0.5.1
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pbr==5.4.5
pycparser==2.20
pylint==2.5.2
python-telegram-bot==12.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
smmap==3.0.4
stevedore==2.0.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
